research-controller/requirements.txt
torch==2.2.0 --extra-index-url https://download.pytorch.org/whl/cpu
transformers==4.51.3
accelerate==0.29.3
bitsandbytes==0.39.0
sentencepiece==0.1.97
peft==0.4.0
# kenlm==0.0.0  # Skipping for now due to build issues, will install later
openai==0.27.7
datasets==2.12.0
wandb==0.15.5
tensorboard==2.13.0
rich==13.3.5
pyyaml==6.0
pytest==7.4.0
scikit-learn==1.3.0 